make use of fmu modelDescriptions wip
