+++ dsl/src/main/kotlin/no/ntnu/ihb/sspgen/dsl/Scoped.kt
@@ -1,0 +1,5 @@
+package no.ntnu.ihb.sspgen.dsl
+
+@DslMarker
+@Target(AnnotationTarget.CLASS, AnnotationTarget.TYPE)
+annotation class Scoped
